creation des api d'ajout de profil propriétaire, du champ photo dans driver

+++ backend-spring-boot/src/main/java/com/example/backend_spring_boot/repo/VehicleRepository.java
@@ -3,5 +3,5 @@ package com.example.backend_spring_boot.repo;
 import com.example.backend_spring_boot.model.Vehicle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleRepository extends JpaRepository<Vehicle, Integer> {
+public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
 }
